tgcw_zhaobiao/requirements.txt
beautifulsoup4==4.11.2
PyYAML==6.0
requests==2.28.1
SQLAlchemy==2.0.4
